Finish abstracting the firebase storage engine
